Fixed issue 35688: It is not possible to delete a record in tree view
The problem was related with Pre Delete Client Side Event Handler. The problem was
that a isc.clone method didn't work as expected. isc.clone method create a deep
clone of an object that can be edited without affecting the original but does not
handle looping references (will infinite loop). For this reason, in case of clone
currentGrid.getSelection() object in tree view an error was thrown.

In order to fix this problem, isc.shallowClone is used instead of isc.clone. This
method handle looping references properly.