[login-perf] Open the registration pop-up on Workspace creation
Due to the changes to avoid an extra request to determine whether a pop-up (registration, heartbeat, etc.) must be shown or not, the pop-up was being opened a bit earlier causing that it was not possible to close it in Firefox. Now the opening has been postponed to the moment where the Workspace tab is initialized.