fixes bug 35771: Can't login with user having a role without access to any org
There was an error on the query launched to retrieve the list of warehouses that can be accessed by a role using the Profile navigation bar widget. The query was failing if the role did not have access to any organization.

We do not need to execute the query on that case, so we now directly return an empty map.