Fixes issue 35772: Orders with Invoice Terms different from Customer Schedule After Delivery cannot be invoiced if the bp has an invoice schedule
Sales Orders with Invoice Terms different from Customer Schedule After Delivery
(Immediate, After Delivery, After Order Delivered) cannot be invoiced if the business
partner has an invoice schedule defined and the order does not fill its requirement

The problem is that the order is not returned by the C_INVOICE_CANDIDATE_V view.
To avoid it was fixed the condition that takes into account the invoice rule of
the order and the sales invoice frequency.