Fixes issue 35636: Business partner selector get stuck in create lines from invoice in hgvol environment
In high volume environment, when the Business Partner selector is opened from Create
lines from process in the Invoice window, it takes too long to load results and as
consequence the browser was stocked.

To improve the performance in this selector, was changed the BusinessPartner_data.xsql
methods to use UPPER instead of C_IGNORE_ACCENT function, this way it uses the index
created in issue 35210 and improves of performance it introduces.

--HG--
extra : rebase_source : f7cb1ed02f48db8b3dc7208bd0064613c1088b86