Fixed issue 35805: Fields redraw instantly in grid under some circumstances
The problem was that numeric fields redraws instantly in grid view when a numeric field
was referenced by a read only logic. This reported problem was caused by the partial fix
of #32202 regression.

The fix of #32202 was partial because the added condition did not taken into account the
grid view properly. This problem has been resolved by take into account form and grid view.

Besides another problem was found and resolved with this fix. Since this regression was
introduced, the #23136 issue was broken too. This problem has been resolved too with this solution.