fixes issue 35769: Firefox browser freezed sometimes after showing some pop-ups
After the changes introduced in project [1], there was a race condition reproducible in Firefox (not in Chrome).

The heartbeat/registration pop-ups are now displayed before the moment they were shown. Together with this, when generating the content of the pop-up, the standard document.write() is used since the very beginning, and according to the documentation[2] it is recommended to call document.close() to tell the browser to finish loading the page.

The document.close() was not being used and this seems to be the root cause of the race condition: by adding this function Firefox is not getting blocked and the instabilities in CI are gone.

[1] http://wiki.openbravo.com/wiki/Projects:Login_Process_Perfomance_Improvements
[2] https://developer.mozilla.org/en-US/docs/Web/API/Document/write

--HG--
extra : rebase_source : b383bca6108409f79aebbd71502fe2122bc7ba91