fixes bug 35825: Default org/warehouse not sorted properly in profile widget
After the changes introduced on issue #35553, the content of the profile widget fields (role, organization, warehouse) is sorted in the client side. The sorting was already being done when displaying the pick lists, but the value map used by those fields was not being sorted.

The default value of the org/warehouse fields when changing the role is the first value of their respective value maps. For this reason, since now the value maps for those fields is always sorted: the OB.User.userInfo object is instantiated with the mentioned value maps sorted by identifier, thanks to the call to 'sortByProperty' added in the user-info.js.ftl template.

--HG--
extra : amend_source : a216c77470b3217149523f8cb07cc4aa43f0bdfd