[login-static] Do not need to keep mBeanServer as an instance variable
On the first call to ManagementFactory.getPlatformMBeanServer() method, it creates the platform MBeanServer. In subsequent calls, it will simply return the initially created platform MBeanServer. For this reason, we do not need to keep an instance variable for the MBeanServer in OBManagementFactory and we can use an static method to register the MBeans.