Fixes issue 35682: Return From Customer pick and edit is slow with big data
With big data, the query executed during Pick & Edit process, in Return From Customer
window, the response time is high, almost 3 minutes.

To improve the performance was redefined the query in ReturnFromCustomerHQLInserter.java
to use the NOT EXISTS function instead of use of SELECT query to only see if doesn't
exists lines with discounts. It also avoid the translation hibernate was doing to a
query with join between c_orderline and c_order_discount tables, and avoid iterate
all c_orderline records if any exists with a discount related.