Fixed issue 35877: On change functions not executed for numeric parameters
The problem was that on change functions was not executed for numeric parameters.
This problem was caused by the partial fix of #35658 regression.

The fix of #35658 was partial because the added condition did not taken into account
numeric fields in P&E windows which are out of a parameter grid. This problem has
been resolved by take into account this case.