Fixed issue 35954: Fetching data when it is sorting by a null FK is not working
Fetching data when it was sorting by a null foreign key was not retrieving any record when should
had been retrieved data.

When AdvancedQueryBuilder class is created order by clause and it is working in resolve joins a
canUseInnerJoin method is invoked. This method is used to check if the left join can be safely
replaced with inner join when it is building order by clause. This method is not take into account
all the posibilities.

The problem has been resolved by take into account when any property of the list is not mandatory
and is not primitive. In this case left join can not be replaced with inner join. The rest of the
cases can be safely replaced it.