Fixed issue 35512: Payments are not correctly loaded in Web POS.
When loading an order in Web POS, a payment is generated for each payment schedule detail in the order. This is not correct. A payment must be loaded for each distinct payment related to the payment schedule details of the order, summing the quantity of the details that belongs to the same payment.