fixes issue 35966: Custom skins for the login page are not working
After the changes introduced on issue #35386, the loginStyles.css file is requested to load all the styles required for the login page. Since then, as this new file is not included in the modules that provides a new skin, the styles could not be loaded properly.

To avoid this problem in a transparent way, when compiling the loginStyles.css is copied together with the skin resources in the proper location. This file will be a copy of the Openbravo_ERP_250.css file provided by the skin itself.