Fixed issue 35674: Concurrency problems applying datasets
The problem occured applying the same dataset (at the same time) for the same client,
organization and version. This problem is not reproduced always. It is a concurrency
problem not managed properly.

In order to resolved this problem two diferent actions have been done:
- Added a unique constraint in ad_orgmodule in order to prevent duplicated data for the
same client, org, module and version. This table is used to check when a dataset is
applied. Besides a BuildValidation is added in order to prevent an error when the unique
constraint applied.
- The same connection is used when applying datasets in order to ensures that all the
modifications in the database are done in the same transaction. Besides some checks has
been improvement.

Now the problem has been fixed by take into account concurrency problems properly.