[myob-request] Improve queries to retrieve widget instances
The widget instances with widgets allowing anonymous access were not being properly retrieved when the role had not explicit access to the widget. That was happening because the query was not correct. Now the available widget classes list is calculated separately and used as part of the filter to retrieve the widget instances available for the current context.