Fixed issue 35760:Can not create computed columns in ad_org or ad_client tables
The problem was that it was not possible to create and export a computed column in ad_org and ad_client tables.
This problem was because computed columns of organization entity should have not organization as a property and
it should have not implement OrganizationEnabled. At the same time, computed columns of client entity should have
not client as a property and it should have not implement ClientEnabled.

In order to fix this problem two changes have been done:
- Now generate entities task is take into account when a computed column is created for Organization or Client
entities. Besides, when the computed column class is created, entityComputedColumns.ftl has the needed logic to
decided if implements ClientEnabled, OrganizationEnabled or both.
- When the Hibernate mapping for the tables and entities are created (DalMappingGenerator),
generateComputedColumnsClassMapping method is take into account when the computed column is for Organization
entity or Client entity.

It is now possible to create a computed column in the ad_org and ad_client tables.