Fixed issue 35842: Process scheduling data can be set in an inconsistent state
The problem was that it was possible to edit scheduling settings for an already scheduled process
request, leaving it in an inconsistent state because the configuration seen in the UI is not the
one that is actually in use.

In order to resolve this problem all fields affecting scheduling settings are read only when the
process request is in scheduled or processing status.