Fixes issue 34432: Accounting information not grouped by invoice when posting
Accounting information was not properly grouped when posting a payment related
with one invoice created from several orders. Accounting info was splitted by
order when this is not necessary and confusing for end user.

Now if there is a previous Payment Detail that belongs to the same Invoice, the amounts
are added.

--HG--
extra : rebase_source : e9b3b1ca6538f1da94912a88962280830db275de