Fixes issue 35832: Multithread support in import/export sample data
Now the processes that import/export sample data support multithread. To do so, the max.threads parameter must be passed to the ant task. If the parameter is not passed it will use half the available CPUs. For instance, to use 4 threads to import the sample data in an install source, this command should be used:

  ant install.source -Dmax.threads=4

--HG--
extra : rebase_source : 18c7e485fbb6a4a562f42ee9da09231a973b316a