Fixes issue 36099: Price Correction Background finishes with error if a goods receipt is invoiced after the costing migration process was run
Price Correction Background finishes with error if a goods receipt is invoiced
after the costing migration process was run. The problem is that the transactions
previous to the migration are generated without m_costing_algorithm_id and the
c_invoice_post function in this case, sets the transaction as checkpricedifference = Y.

To fix that, transactions previous to the migration will not be adjusted.
To avoid NPE in PriceDifferenceProcess.calculateTransactionPriceDifferenceLogic
was included a validation taking into account all explained above.