Fixes issue 36180: Add Multiple payments pick and execute performance
When the BP filter combo box of Add multiple payments P&E the resultant query is
slow. The problem is related to the usage of AD_ISORGINCLUDED function.

To avoid it, and improve the performance, was used an EXISTS clause in the AD_ORG_TREE
table instead of using AD_ISORGINCLUDED function.