Fixed issue 36109: Delete Client process doesn't managed exceptions properly
The problem was that Delete Client process didn't manage exceptions properly. If Delete Client Process
was fail, it did not show any error message and leaved the triggers disabled.For example, if a problem
in the database was raised, the UI was not take into account the problem as expected and the user did
not realize anything.

Now the process does handle exceptions as expected. If the Delete Client process fails, all the
constraints are enabled again. This occurres because enableConstraints method is executed. Now a
messagge is displayed in the UI in order to inform to the user. Besides, if the Delete Client process
fails a rollback is done.