Fixes issue 35764: The browser get stucked when printing the receivable aging schedule report showing details and having a lot of invoices pending.
After printing the receivable/payable aging schedule report on an environment
with a lot of invoices pending to be paid, the report took too long to be printed.
Most of the time was consumed by chrome rendering the server response.

In order to reduce the memory usage, was limited the displayed data with the value
of the "ReportsLimit" preference.

When record count exceeds this limit, is shown a warning informing data has been
limited. As problem happens when the report runs in HTML format, data only will be
limited in this case and not when report runs in PDF or XLS formats.

As Aging Schedule reports are defined as Process Definition, and results of run as
HTML is shown in a new window containing the JR with results of the
selected filters, was added to the affected JRXML files a new hidden Text
Field that will be shown only when the data is limited. This way, user will be
notified after executing the report in both windows, selected filters and report results.

To achieve this functionality the following changes were done:
-Adapted AgingDao_data and AgingDao to make possible to limit queries to PG and
 Oracle databases.
-Adapted AgingBalanceReportActionHandler to limit data if report is created in
 HTML format. If the data is limited are stored as a session variables to make
 it available to AgingDao methods. Then a warning message will be shown in the
 affected reports if the limit is reached.
-Added a new "warning" parameter to AgingScheduleDetailHTML and AgingScheduleHTML
 reports. It is linked to a hidden text that only will be displayed if the maximum
 of records count defined in the preference to reports is reached.

--HG--
extra : rebase_source : 9a6c8baddb56b7fb34eb36cc71c978d211b07c63