Related to issue 36061: Code review improvements
Avoid NPE when calculating the cost of a production which consumes a non stocked product, as non stocked product will not have a related M_Transaction.
For taking the cost of this non stocked transaction, we should look for a standard cost, same as we do in DocLine_Material.getProductCosts().