Fixed issue 36216: servlet callout SL_CreateLines_Conversion_UOM is not working
The reported problem is the same that it had been fixed in this issue #35540 . In this case, SL_CreateLines_Conversion_UOM callout is still required to be manually executed as a servlet.

Callouts should only be used in generated windows, for manual old 2.50 ones, as this case, the mapping is not generated by defining it as a manual map.