Fixes issue 36377: Cannot launch Valued Stock Report
Cannot launch Valued Stock Report if the role does not have
access to M_COSTING_RULE and M_TRANSACTION tables.

The Report access to M_COSTING_RULE and M_TRANSACTION tables and if the user role
does not have access an exception is raise.

Now the context set the admin mode when the report try to access the tables and
restoring the previous mode when finish the access