Fixes issue 36412: Orphan Line removed from Return To Vendor
The orphan lines was removed when a Pick/Edit is open because if the selected Lines
from the Pick/Edit is 0 all lines are removed and the methods ends.

A refactor was made to take into account the selected orphan lines in the
selected lines process, also a null save validation was done with the goods shipment
line because an orphan lines doesn't have a goods shipment line.