Fixes issue 36379: Org Access is not added for inactive orgs creating new Role
When a new Role was created, all org access are given for the existing organizations,
but not for those inactive.

It was caused because OBCriteria applies filter active='Y' by default.

To fix it, criteria is set to not filter by Active organizations, this way it also takes
into account those inactive organizations.