[250removal] remove ExecuteQuery.executeUpdate & selectAndSearch
ExecuteQuery is a helper class used very seldomly to manually run sql queries
without using xsql. The executeStament part to be able to run i.e. DELETE was
only used by the DataGrid removed in the previous commit.
Apart the selectAndSearch method was a special case method only used in the
generated servlet for the old 2.50 windows which have been removed already.

This commit removes both methods and only leaves the basic 'select' method as
it is still used in a few other places.

Checking in getAllModules did not find any module using the ExecuteQuery class
at all.