[250cleanup] Disable/remove Buscador from 2.50 ToolBar
Similar to previous commit for ExportGrid. 2.50 ToolBar has a button to
use the 2.50 Search Popup for the old generated windows.

The only prepare* methods not removing this button are:
- prepareEditionTemplate
- prepareRelationTemplate
- prepareSimpleToolBarTemplateFrame

Out of those checking getAllModules only found 1 hit for the first in the
org.openbravo.deprecated.cleanupv3 module into which many years ago old
code was moved when it was deleted from pi.
That InvoiceVendorMultiLine.java was very old grid edit functionality when
2.50 grid was not yet editable. As that is related to removed 2.50 grid it
will no longer work anyway so is not a stopper here.

Same as before the SEARCH + SEARCH_FILTERED buttons are disabled and when some
module should trigger a code-path using them -> An error with stacktrace will
be raised to easily find such code.