[250cleanup] Remove flush from UIReference constructor.
Previous code-cleanup allows for deletion of now unused member conn
which is initialized in constructor.

However as that previous init was done doing an inplicit flush this is
slight behavior change which is why this change is done in separate commit.

If any code should rely on that implicit flush being done that would be
definitely wrong and removing the flush in useful as it is unexpected by
calling code and can affect performance in some cases.