[build-cleanup] Remove ant javadoc targets from src-core,src-trl,src-wad
This removes the following ant calls from main build.xml file:
- core.docs, wad.docs, trl.docs

Javadoc for all three items is included in main ant generate.java.doc target
already. src-wad + src-trl as very tied to Openbravo so not useful outside
of Openbravo project. src-core in theory could be used outside, but that will
be very hard, was never an official feature, and no known user is known.

Also each of the 3 target had matching eclipse .launch files which are also
removed as not useful with the ant targets.

Note:
The rm call to remove the docs subfolder where javadoc would be generated is
left in the files to cleanup any earlier created folder or javadoc.

--HG--
extra : rebase_source : 88dec903458ed8863455131319547a185606f27f