Related to issue 36398: Hide secondUOM fields if product doesn't have secondUOM
When a product with a second UOM was selected the associated fields are shown,
but not hidden when the product is change and doesn't have a second UOM because
the has_second_uom var is not updated with the corresponding value according
to the product selection.

The value of has_second_uom var is updated according to the product selection
in the result for a correct use in display logic function.

--HG--
extra : transplant_source : C%BA%9A%3A%D9%E9%E1%9C%AF%FAd%1D%FE3%92w%5E%D4%DA1