Fixed issue 36522: With Synchronize mode the cashup could be incorrect
The solution is to sort the array of receipts from the webPOS in the same way
as it’s retrieve in the websql database of the client.

The cashup is an accumulative sum of all transactions from webPOS. It’s
important to the order of the transactions generated to avoid differences
problems in the cashup.

This problem only happens in synchronize mode, where the cashup is
always retrieve from the backend instead of using the cashup report
from the client.

--HG--
extra : rebase_source : b8cf96443f7d92afca4d9771dc7d8dc4a6ad1568