Fixes issue 36455: Cannot complete a shipment if related to a draft reservation
When a reservation was removed from Manage Reservation process in Sales Order,
the reservation keeps the association with the line and Not support operation
exception was raised because the reservation changes from DR to CL.

Now the Manage Reservation process in Sales Order removes the reservation
if the reservation has no lines.

Also the c_orderline2_trg only closes reservations in CO and HO status.