Fixes issue 33397: Reactivate a reservation doesn't work in Oracle
When you try to reactivate a reservation that is not linked to an order, the process
fails with the error message. Problem is that procedure is always expecting an order
linked to the reservation, and this doesn't happen always.

--HG--
extra : rebase_source : a560771e6947c9029b36098f813f0b028d01525e