Fixes issue 36659: Closing inventories are not leaving stock valuation to 0
When Costing Migration Process is executed closing inventories are not leaving
stock valuation to 0. In calculateCosts function of CostingMigrationProcess class,
in order to adjust the cost of the last transaction, the movementqty was wrongly
compared with the stock because it doesn't take into account that movementqty will
be negative (as it belongs to closing inventory transactions) when stock is positive.

To fix it, the movement quantity is negated when comparing with stock.