Fixes issue 36670: Process failed during execution when posting reconciliation
When does not exist currency conversion a "Process failed during execution" message
was shown instead of "Document currency cannot be converted to the Accounting Schema
currency." during execution when posting a reconciliation.

Problem was that FactLine.convertByRate() and FactLine.convert() methods return was not checked when calling them from FactLine.createLine() method.
Now, when they return false, document is set to not convertible status.