Fixes issue 36207: EntityPersistenceEventObserver validations has high log
Entity validations created extending EntityPersistenceEventObserver, when the
validation fails an exception is thrown, and it is registered in the openbravo
log file. These kind of exceptions are functional validations and should not be
registered in the openbravo log with a high level.

The problem is some of observers are throwing OBException, which by default log
stack traces.

Since 3.0PR17Q4 OBExceptions doesn't logs exceptions with stack trace by default
when a new OBException is thrown, this fix specify the second parameter of
OBException constructor as true to just logs not functional validations.

--HG--
extra : rebase_source : 00596ae8608e98d1c6b450b04aae292a91acbd77