Fixes issue 35427: Conversion of Valuation not correct in Pareto Product Report
The conversion of the currencies in Pareto Product Report is not the same as in the
Valued Stock Report. The problem is that the total valuation is converted to the
currency for today, instead of converting each Transaction in the day it happened.

This fix uses the transaction's date instead of the current date to convert
the valuation. Also is grouped by transaction's accounting date to have an homegeneous
behaviour like in CostingUtils.getCurrentValuedStock() and CostAdjustmentUtils.
getValuedStockOnTransactionDate() methods.

--HG--
extra : rebase_source : 0960a134bb36c0832a625bd807d45b503d19baca