fixes bug 36459: error when refresh after discard changes before FIC response
This problem was solved with the code that fixes issue #31434 just for the case of cancelling a record which does not contain any change. Otherwise, the issue could be still reproduced because of the fact of having an if statement with two code branches inside the discardEdits function. That code was forgotten to be included also in the first branch.

That if statement was created in order to identify if the record being cancelled contained changes. In that case, a dialog was being shown to the user in order to ask for confirmation to discard the changes.

After some time, that dialog was removed (see revision 92ecb7ac760a). That means that since that point the if statement is no longer needed. For this reason, the two branches have been merged.

Note that this change also fixes completely issues #31434 and #33626, as they were partially fixed (fix only present in one of the branches).

--HG--
extra : amend_source : 0e822ce2484e5635cc80c7954bbe3ffc71671212