fixes issue 36683: CSS minifier breaks @font-face directives
CSSMinimizer have problems dealing with selectors like url('data:image/....'). The url content need to be surrounded with double quotes in order to minimize it properly. For this reason a regular expression is used in order to replace the single quotes of the url definition with double quotes.

The problem here was that if additional single quotes appeared after the url definition, like for example:

  src: url('./../font/Roboto-Condensed/Roboto-Condensed-Bold.woff') format('woff');

the replacement was not being done properly.

To fix the problem the regular expression has been changed in order to look for the closing quote of the url content. Thus, we ensure that we are replacing just the single quotes of the url value.

--HG--
extra : rebase_source : 64c6d1f80d92aa66e7d957022829e7c6fda5f3da