Fixes issue 34915: Dimensional Reports does not show values if Org is empty
When Dimensional Reports were executed did not show any value if Organization field
was empty. When no organization was selected the session's organization and it childs
were taken by default to filter results. It implies that if you had selected a non
legal organization in the session, then reports were ignoring records from it parent
and brothers.

To set homogeneous the behaviour in all these reports, the organization field has been
set as required and different than blank.
When any organization is selected then an error message is shown explaining it.