Fixes issue 36681: Insufficient stock in Create Standards Process of Work Effort
The problem was that in M_GET_STOCK_PARAM function v_Quantity variable does not
accept decimals and when quantity is decreased on iterations decimals are ignored
and the next records from M_STOCK_PROPOSED table are deleted.

To fix that, the v_Quantity variable has been declared as NUMBER to accept decimals.