fixes bug 36716: Tree view stops working when returning back from form view
When opening the form view of a tree view node, Smartclient copies the values of the node properties into the form but skipping the properties involved in the tree management (these properties the child nodes are also not copied).

When returning back from the form view, the tree node is refreshed with the values of the properties stored in the form. This caused the lost of the tree information of the child nodes because, as mentioned before, the tree properties were not copied when opening the form view.

Apart from the skipped properties, the tree information of the child nodes should not have changed in form view. Therefore, to fix the problem now we do not override the children information when returning from the form view.

--HG--
extra : rebase_source : e7a9919236618fb3cfe21c86d8775043d193e6df