fixes bug 36739: Can't use fields based on some properties in selector picklist
With the fix for issue #36268 the selector fields based on derived properties that are shown in the picklist are sent as 'extra properties' in order to calculate them properly.

The problem is that the fix was not complete when the derived properties were not primitive. In that case the DataToJsonConverter class was not calculating properly the referenced property.

Now that kind of properties are calculated properly.

--HG--
extra : amend_source : 9d750faa5a8aacd7b89e6e6aad46cbc6f1649aa2