Fixed issue 36598: Multirecord Process sets only one organization in parameters
A standard process can be defined as multi record process to be able to execute it for more than one
record. In this case it can be selected records belonging to different organizations. The reported
problem was that in this situation it was take into account only the first organization sending it
as "_org" parameter.


This problem has been fixed by managing this situation properly. Now, when a multirecord process is
launched and more than one record is selected, the natural tree of writable organizations of the
current context is calculated instead of the natural tree of the first organization of the first
selected record.

--HG--
extra : rebase_source : eac4dae75b334398c4aa2057c15883821f10c946