Related with the issue 36598: now selectors are working as expected in P&E
Selecting multiple records and opening the process with a selector in the 'P&E' it is
failing.

The problem has been fixed by take into account this case. Now when the selector is open
shows records in the tree of all orgs current role has access instead of records in the
org tree of the last selected record.