Fixes issue 36690: Insufficient stock in Create Standards Process of Work Effort
When Create Standards Process of Work Effort - Production Run tab was executed
if stock is in different priority warehouses, an error message Insufficient stock
is shown even if there is stock of the product.

Problem was that in M_GET_STOCK_PARAM function v_qtyaux is decreased on each loop
and v_finalqty is taking into account stock in warehouses of the same priority than
currently analyzed, when v_qtyaux is updated always is subtracted the sum of
m_stock_proposed lines with priority less or equal than current and it is wrong.

Solution is always compare the requested quantity (v_quantity) with sum of quantities
from m_stock_proposed lines with priority less or equal than current iteration.