Fixes issue 36705: More than one pending payment plan detail for same invoice payment plan
When the payment schedule details were generated it was comparing the invoiced
amount of related order with the sum of lineNetAmt + taxAmt of it lines, as tax
calculation can be done at document level, and this sum is done at line level it
is possible that exists differences between this sum and totals already rounded
in the invoice.

Also was considered an order completely invoiced if it dosn't has lines with ordered
quatities different than invoiced. When exists orders with lines with lineNetAmt = 0
it was taking it as completed invoiced and then the whole amount of the psd was included
in the invoice and it generates a difference between the total of document and the sum of the
lines that was created as a new payment schedule detail.

To avoid this situation, only are considered order lines with lineNetAmt different than zero
to use the whole amount of the psd or to re-calculate the correct amounts from Payment
Schedule and Payment Schedule Details related to the order and invoices.