Fixes issue 36707: Price List Schema is not working in a "hierarchical" way
Price List Schema with several lines was not working in a "hierarchical" way,
because every time it was processing a new M_DiscountSchemaLine it was deleting
the previously calculated product prices, and it always save the last calculation
taking the last line of the price list version as basis.

To fix that, only are removed previously calculated product prices that will be
regenerated at start of the process for the Price List Version being processed.
And these product prices only are removed and inserted once, the first time the
discount will be applied. For next discounts, product prices will be updated using
previously calculated prices.

This way, apply several lines of the Price List Schema will be executed sequencially:
first is applied the discounts for the first line and all prices calculated will
be the basis for the next one to apply, this process until all lines be applied.