Fixed issue 36611: Selection is lost when an error is raised removing a record
When were removing a row and an error was raised (e.g. Deleting a already processed "Good Shipment" is stopped
by a trigger) the selection of the record was lost. This generates an inconsistent grid state that creates all
the reported problems.

The problem is fixed by take into account the described situation removing the deselectAllRecords method because
it has no sense to remove the selection if it is not possible to know if the removed process ends succesfully.
Now the selection is maintained if the record can't be removed properly.

Now the selection is maintained if the record can't be removed.