Fixes issue 36725: Check role write access to organization of parent record
If the current role does not have writable access to the record selected in a tab, the user should not be allowed to create records in its subtabs.

Now this is taken into account in order to enable/disable the toolbar buttons that create new records, and also in the logic that creates a link to create new records in a tab if the gr
id is currently empty.

Now it will not be possible to enter records in a subtab if:
- The subtab has an editable organization field
- The user's role does not have writable access to the organization of the record selected in the parent tab.

Now the list of the current role of the user writable organizations is available in OB.User.writableOrganizations.

--HG--
extra : rebase_source : fdc3d7cce198d0682fc4e2c66da636b3e8eaa139