Fixed issue 36112: Checksum is updated properly on install source task
The problem was that checksum was not updated on install source task. The problem was
in ImportSampledata because this task can execute module scripts that modifies the
database. (e.g. create or modified a table).

The problem has been resolved by take into account this situation. Now the checksum is
calculated after the modules script are executed.