Fixed issue 30489: Templates hierarchy are not applied as expected
The problem was that hierarchy was calculated wrongly when it should be applied the
configScript files. The generateTemplatesTree generate a wrongly list of templates
in terms of sorting. It means that templates were properly located but applied in
the wrong order.

In order to fix this problem, getSortedTemplates method is used to retrieves the
sorted template list. Now, the configScript files are applied properly, it means,
the hierarchy is applied as expected. For example, the changes defined in the
configScript file of the retail pack are applied after the core ones.