Fixed issue 36371: Delete payment schedule details is slow when paying a layaway
Previously:
  When creating a payment for a layaway, it create a new fin_payment_scheduledatail
for the payment, then delete all the non linked fin_payment_scheduledatail's and then
create a new one for the remaining amount.

After the fix:
  Instead create a new one for the payment search for any non linked, if found reuse it,
and create a new one for the remaining.

Note: the non linked are still deleted before create the one for the ramaining, but
in normal case should not be any to delete. Only will be when creating the order in
the erp and seting many schedule details, webpos always create only one scheudle detail.

Also needed to remove the set of the id when creating the remaining one, since it fails
with the new changes and they are not needed, since the correlation of ids between the client
and the backend is not needed for the case fin_payment_scheduledatail_id.

--HG--
extra : rebase_source : 8fd800729a1eb93e3581a8a4991741516c6ff9fe